requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
boto3==1.34.55
botocore==1.34.55
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
distlib==0.3.8
Django==5.0.2
django-filter==24.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
identify==2.5.35
iniconfig==2.0.0
jmespath==1.0.1
minio==7.2.5
nodeenv==1.8.0
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.2
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.0.7
virtualenv==20.25.1
